+++ firebase-abt/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Unreleased
 
 
+# 23.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+* [removed] **Breaking Change**: Stopped releasing the deprecated Kotlin extensions
+  (KTX) module and removed it from the Firebase Android BoM. Instead, use the KTX APIs
+  from the main module. For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration).
+
 # 22.0.0
 * [changed] Bump internal dependencies