+++ firebase-dataconnect/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
 
 
+# 17.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+* [changed] Removed superfluous and noisy debug logging of operation variables.
+
 # 16.0.3
 * [fixed] Fixed occasional `NullPointerException` when registering with
   FirebaseAuth, leading to erroneous UNAUTHENTICATED exceptions.