+++ firebase-datatransport/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
+
+
+# 20.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+
+# 19.0.0
 * [changed] Bump internal dependencies.
 
 # 18.2.0
 * [feature] Expose the transport factory as a qualified dependency based on the endpoint connected.
+