+++ firebase-messaging-directboot/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 
+# 25.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+
 # 24.1.2
 * [unchanged] Updated to keep messaging SDK versions aligned.