+++ firebase-perf/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
+* [fixed] Fixed an ANR on app launch. [#4831]
 
+# 22.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+* [removed] **Breaking Change**: Stopped releasing the deprecated Kotlin extensions
+  (KTX) module and removed it from the Firebase Android BoM. Instead, use the KTX APIs
+  from the main module. For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration).
 
 # 21.0.5
 * [changed] Updated `protolite-well-known-types` dependency to v18.0.1 [#6716]