+++ firebase-storage/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Unreleased
 
 
+# 22.0.0
+* [changed] **Breaking Change**: Updated minSdkVersion to API level 23 or higher.
+* [removed] **Breaking Change**: Stopped releasing the deprecated Kotlin extensions
+  (KTX) module and removed it from the Firebase Android BoM. Instead, use the KTX APIs
+  from the main module. For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration).
+
 # 21.0.2
 * [fixed] Fixed an issue where tests were depending on a deprecated feature of the test framework. (#6927)